requirements.txt
Python 3.8.3

Django==3.1.1
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
drf-yasg==1.17.1
drf-yasg-examples==0.1.1